cdk/requirements.txt
aws-cdk.core==1.46.0
aws-cdk.aws-ec2==1.46.0
aws-cdk.aws-rds==1.46.0
aws-cdk.aws-secretsmanager==1.46.0